Fix the folder path for recreating a konnector folder
When a konnector is executed, the stack ensure it has a folder where it
can be put its files if needed. The path for this folder is taken from
the account, but if it is missing, a fallback will use the konnector and
account data. It was using the slug of the konnector, and it will now be
using the konnector name instead.