Allow to manually certify as flagship an app
If an OAuth client send the user to the authorize page with a '*' scope,
and it has not been certified as the flagship app, an email is sent to
the user with a 6-digits code. The code can then type this code to
confirm that they want to use this application.