fix: Return the real error for /capabilities/settings (#3307)
We were returning the same error message for all
the possible errors.

Now we return the specific error.

We find out this issue when requesting this route
with an expired OAuth Token. And since the only
message we had was "Forbidden", the refresh process
was not launched.

Thought, we had the right message in the WWW-Authenticate
header.