Optimize listing of files (#3336)
When a request is made for listing files (showing a directory or find
request), and the response contains many images or PDFs, we were
creating the secrets for the thumbnail links one by one. When the stack
is configured to use redis, the latency for each call is added, and it
can take up to a few seconds for hundreds of documents. We are now
batching the calls, which should speed those requests.