Fix JSON serialization of app manifests (#3377)
When a webapp or konnector is installed/updated/removed, and we have
several websockets listening to changes for io.cozy.webapps or
io.cozy.konnectors, the stack try to serialize to JSON the same manifest
several times concurrently. Golang doesn't like that a map is read and
written concurrently and can crash. The fix is to clone the map to have
one map per serialization.