Fix accepting sharings (#3396)
When a user accepts a sharing, there are multiple processes in parallel.
The stack for the receiver Cozy sends a message to the owner's Cozy to
tell it that the sharing is accepted, then setup the sharing. The stack
for the owner's Cozy is doing stuff, and then starts the replication.
So, we can have two goroutines in the receiver side that are trying to
create the sharing folder. We are trying to mitigate issues with the
eventual consistency of CouchDB.