Fix the BI reconnect flow
The BI reconnect flow is similar to the OAuth flow for konnectors: it is
meant to be triggered via a browser opening a tab, not via a fetch/xhr
request. As such, there is no token for checking permissions, but we can
rely on the session to ensure the security.