Forbid patching md5sum by clients
The PATCH method for io.cozy.files allowed to update the md5sum. It was
introduced for fixing missing md5sums. It looks like this endpoint has
been misused by some clients and can create FS errors (content
mismatch). So, let's remove this possibility and the fix command that is
no longer used.