Avoid extracting autofilled passwords
If a user has shared a folder in read/write mode, someone can put an
html page in this folder, creates a download link, and sends it to the
owner of the Cozy. Password managers can autofill the password on this
html page, as it is the same domain as the Cozy authentication pages.