Fix redirect when cancelling OAuth for a konnector
For Google and BI webviews, an OAuth danse is made, and if the user
clicks to cancel it, they are redirected to the stack. We want to go
back to the app where they started the OAuth danse, not to an hard-coded
slug (store), because the store app may be not available and it makes
harder to have a coherent User Experience.