Allow to use big payload for webhook triggers (#3416)
The payload from webhooks was read from the COZY_PAYLOAD env variable by
the konnector or service. But, env variables have a size limit (defined
by the linux kernel), which makes impossible to use large payload. Now,
when the stack sees a big payload, it will write it to a temporary file
and will just put @ + the filename of this file in COZY_PAYLOAD.