Fix encoding issues of dir/file name in CouchDB
CouchDB has changed the way several documents with the same field in
different encoding are returned in view requests. But, it wasn't what I
was expecting, and with CouchDB 3.2.1+, it wasn't possible to change a
file or directory name to a new name when just the encoding has changed
(eg NFC -> NFD).

Cf https://github.com/apache/couchdb/issues/3773