Fix trigger creation for konnectors (folder_to_save) (#3507)
When a trigger is created for a konnectors via the stack endpoint
POST /jobs/:slug/trigger, the stack tries to mimick the behavior of
harvest. In particular, it needs to to fill if the folder_to_save field
of the trigger if the konnector manipulates files. Harvest looks at the
manifest of the konnector for the fields.advanced_fields.folderPath (and
the stack was doing it too) and folders (which wasn't done by the stack
before this commit).