Fix race condition on concurrent uploads (#3517)
When there are 2 concurrent uploads for the same file, we may have a
race condition on the candidate version. The candidate will twice be
the same version, and we need to avoid keeping it for one upload and
cleaning it for the other, as it creates a FS inconsistencies: the
version will exist in CouchDB but not in Swift.