Fix race condition on app updates (#3527)
When an app is asked to be updated several times at the same time, it
may happen that triggers for the services will be messed. This is fixed
by adding a lock for updates.

The happy path (when the app is already up-to-date) is kept the same. It
is only when an update is attempted that a lock is acquired. As such,
the app document must be reloaded from the database as it may have been
changed by another goroutine doing the update.