Fix cleaning thumbnails while dissociating files (#3528)
When files with thumbnails are shared, and someone move them outside of
the shared  folder, a copy of those files is put in the trash for the
other members of the sharing. This operation is called dissociating
files, but it didn't removed the thumbnails of the old files because of
an error "nil Context". This is now fixed by setting the context on the
thumbsV2 object.