Make the all_docs endpoint with Fields faster
When requesting all_docs for a lot of io.cozy.files with Fields, the
response can be quite slow (a few seconds). Part of that is CouchDB and
transfering the bytes, and we can't do anything about that, but there is
also another part that is deserializing JSON, keeping only the required
fields, and reserializing to JSON. And it looks like this part is taking
way much longer that we would have expected.

In this commit, we are trying to make things faster by replacing the
full JSON parser by just a tokenizer (and doing some low-level stuff).
I have added unit tests, but the code is tricky, so I'm less confident
than usual on the risk of a regression. And we need to deploy on servers
to really see the performance boost (or lack of).