Do not retry upload on Internal Server Error (#3568)
When uploading a file for a sharing, it may happen that the
io.cozy.shared document is corrupted (bad chain of revisions), and the
upload fails with an Internal Server Error. In that case, it is useless
to retry, as it will fail again, but it is better to skip this file to
allow uploading the next files sooner.