Fix a conflict case for io.cozy.sharings doc (#3567)
When a sharing is revoked, it may happen that a conflict (409 from
CouchDB) is received when updating the io.cozy.sharings document after
deleting its triggers. After that, an error trigger_on_inactive_sharing
is reported by cozy-stack check sharings.

And the fix is quite simple: when we receive a 409, we reload the
document from CouchDB and try again.