Add reference to the sharing directory (#3580)
When a sharing is created for a directory or a file, the stack should
add a reference to the new sharing on the directory/file. It was done
when a recipient accepts the sharing, but it could be too late if the
directory/file was trashed before that: the sharing should have been
revoked.