Fix share-track jobs when they are not in expected order (#3582)
When an io.cozy.file is created and modified just after, it may happen
that the share-track job for the update is processed before the job for
the creation. When this happens, the revs tree for the io.cozy.shared
wasn't correct: the revision 2 was coming before the revision 1. This
commit adds a check to put the revisions in the good order for this
case.