Use interact token for preview URL of sharing (#3583)
* Use interact token for preview URL of sharing

When a user goes to the preview URL for a Cozy to Cozy sharing, and if
their Cozy URL is known (ie they have accepted a sharing), then we
should allow this user to interact with the sharing. For a note, it
means that they will be allowed to edit it. For a folder, they can
upload new files.

* Update model/sharing/sharing.go

Co-authored-by: Erwan Guyader <erwan.guyader@cozycloud.cc>

Co-authored-by: Erwan Guyader <erwan.guyader@cozycloud.cc>