Add a lock around fs.Copier when updating apps (#3625)
When an app is installed or updated, it may happen that several
instances are doing this operation at the same time. There was already a
lock to avoid several installs/updates of an app on the same instance,
but we want to avoid a too large lock, which can create contention.

This commit adds a new lock to avoid FS write operations for
installing/updating an app from several instances at the same time (ie
just after a new version has been published). The happy path avoids the
lock to ensure best performances.

It should help to reduce the pressure on Swift after a new version of an
app is published.