Add the `context.Context` support for `couchdb.BulkForceUpdateDocs`
A `context.TODO()` have been used as a placeholder for all the calls,
even if there is a real context available.

This is due to the fact that the call migration is done programatically
(with `sed`) without taking into account each call environment. This
should be fixed in a following step.