fix: Handle missing port in sharing instance check (#3665)
To check if a sharing member's instance can be found in CouchDB, we
first need to parse the member's URL instance to retrieve its domain.
However, when the Cozy is reachable using a specific port, that port
is part of the instance's domain (i.e. in the form of
`<domain>:<port>`).

We assumed that a missing port in an `http` URL would give port value
of `80` once parsed (and `443` for a `https` URL) but that's not the
case.
Therefore, in these situations, the instance's domain used for the
CouchDB request would have the form `<domain>:` and the instance would
never be found.