Fix the export/import integration test (#3742)
When running the export/import integration test, we are looking at an
email with the export link. For that, we are querying MailHog with the
instance email source+test@localhost. But, it can return us an email for
konnsource+test@localhost from another integration tests. The first
email address is included in the second address, and MailHog matches on
that. The fix consists of having a more specific email address for the
instances of the export/import test.