Add a trigger for saving OO documents
When a document is edited with OnlyOffice, the OnlyOffice server keeps a
cache of the document and will send the document to save to the stack
only if the user asks for a forcesave (ctrl-s shortcut), or after some
time with no tab opened on the document. In our experience, it can be
long and the users can expect the document to be saved to the VFS, and
then synchronized via the sharings / desktop, but OO still has the
document in its cache and doesn't send the save.

So, we decided to add a new trigger that will ask the OO server to
forcesave often to ensure a better synchronisation between OO and the
Cozy.