`client/client.readJSONAPILinks` now take a `io.Reader`
Also move the response body close next to the requests. This move
is required in order to make the `bodyclose` linter happy. It doesn't
handle well the case where the body is closed in a sub function.

In any case this behavior doesn't make the body closing explicite an
can easily lead to missing body closing or double closing.