Create a Redis client based on `/pkg/cache/cache.go` redis logic.
The `/pkg/cache/cache.go.Cache` struct seems to contains two separate
behevior based on the prsence (or not) of a redis client.

If the redis client is present we are using the redis client, if not
we are using an in-memory implementation.

The goal of the next commits is to separate those two logic in order
to avoid to have this `if` condition in every methods.