Use TDD in order to tests all the cache clients the same way
The goal is to tests all the cache implementations exactly the same
way by running the same tests and checking the that the same outputs
are generated.

This allow to easily add a test for all the implementations and to
check that a new implementation have the same behavior than the other
implems.