Create the `Cache` interface and rename the `Cache` struct into `GenericCache`
This is some required work before spliting the two implementation living
into the GenericCache. At the moment the depending if the client in `New`
is nil or not we use the redis backend or an in-memory one.

The goal is to split this client in two separate implementations and
avoid the if condition in each methods.