Do not use `make unit-tests` in `ci` job
The CI goal is to test everything in order to check that all the
codebase is ok. By using `make unit-tests` we are running only a
part of the tests and we are skipping all the time heavy ones.

This review ensure that we are running all the tests and add some
coverage information at the end of the script