Move the vault logic into a new `keyring` package
Previously the vault logic was living inside the `/pkg/config` package
in a global alongs with other config stuff.

This PR goal is to make a clear separation between the vault logic,
where some important secrets are handled and the config logic. It also
remove the global variable and move to the same initialization logic
than the cache or lock package: It create an object and put it into the
`Config` object.