Fix tests for redis scheduler (#3822)
A previous commit has made a refactoring to use a new instance for each
test, instead of using the same test instance for all tests. It was made
to avoid side effects from one test on another, but it has the opposite
effect. The instances were created via setup.GetTestInstance(), which
starts the system broker+scheduler. But the configuration is read when
the first instance is created, and the other tests were inheriting this
configuration. We can avoid those issues by using a custom scheduler in
redis scheduler tests, instead of using the system scheduler.