Fix sending push notification to flagship app (#3823)
If an application is sending a push notification, and the user doesn't
have the mobile app for it, but has the flagship app, the stack will
fallback to send the push notification to the flagship app. But the
firebase configuration can be different for the mobile app and the
flagship app. This commit fixes the firebase configuration used when
fallbacking to the flagship app.