Do not compare PNGs pixel by pixel (#3826)
It seems that there is some output change between the images generated
by the CI and in local. I guess this is due to the different versions
of convert.

We don't want to force a specific version for everyone so we will
only check that the two files a PNGs and they have the same size.