Use the `debugLogger` only with debug log and a domain in debug mode
Previously we were checking if a domain was in debug mode for each
call to `WithDomain`, even if this is for a warning or an error. This
would force almose all the logs to be serialize du to the use of a
sync.Map even if it's better than the previous lock.

In order to reduce this friction, this commit will check if the domain
is in debug mode only when the log have a debug level. In any other case
we log with the classic logger.

It also remove all the logic from the `WithDomain`. This will allow us
to link this method to `Entry` and make it behave like a custom
`WithField` method.