Improve views and indexes update (#3858)
When a Cozy was created before the set of indexes or views for an
instance was last changed, it has a different IndexViewsVersion.
This number can be used by the stack when a Cozy is used to check if the
views and indexes must be updated.

Before this commit, the update process was the same as the creation
process. We iterate on all views and indexes, trying to create them, and
if it fails, look if the view/index exists and has the right definition,
and if necessary update it.

With this commit, the update tries to opmitize this process, by first
looking what views and indexes exists, and only try to create/update it
when missing or not the right definition. It should help to lower the
load on the CouchDB cluster when we release a stack with an increased
IndexViewsVersion.