Fix delegate userinfo endpoint (#3868)
The delegate userinfo endpoint for OIDC was using an Authorization
header for the access_token. But, as it was on the admin endpoint, there
was also an Authorization header for basic auth of admin. And the two
headers were conflicting: they were not manager correctly by cloudery
and stack. We can fix the issue by passing the access_token in the POST
body (as JSON).