Fix a race condition on realtime
When a client asks for UNSUBSCRIBE on a doctype, and just after that for
SUBSCRIBE on the same doctype, there was a race condition where a part
of the UNSUBSCRIBE process where done after the process for the
SUBSCRIBE. It is due to the lock for the mem hub being acquired in
goroutines. We are now getting the lock sooner to avoid this issue.

In practice, this bug was seen when emptying the trash in Drive web and
files were still on the screen as Drive was not receiving the realtime
notifications for them.