Refactor the way `/pkg/logger handle` the debug mode (#3852)
* Use a single logger for all the domains in debug

This commit bring two main advantages:

- It replace the lock on debug loggers by a `sync.Map`. This should
  improve the concurrency for the logs.

- It use a single `debugLogger` for all the logs in debug, instead of
  the previous way.

This review is also a preliminary work to move the logger inside the
`Log` function and checking the `domains` map only when we generate
debug logs.

* Use the `debugLogger` only with debug log and a domain in debug mode

Previously we were checking if a domain was in debug mode for each
call to `WithDomain`, even if this is for a warning or an error. This
would force almose all the logs to be serialize du to the use of a
sync.Map even if it's better than the previous lock.

In order to reduce this friction, this commit will check if the domain
is in debug mode only when the log have a debug level. In any other case
we log with the classic logger.

It also remove all the logic from the `WithDomain`. This will allow us
to link this method to `Entry` and make it behave like a custom
`WithField` method.

* Add the `WithDomain` method to `pkg/logger.Entry`

This will allow to add a package logger with an preset namespace
and manipulate only an `Entry` inside the package.