Fix adding a contact by Cozy URL to a sharing with delegation
If Alice creates a sharing with Bob in read/write mode, and with the
open_sharing set tot true, Bob can add recipients to this sharing. If
Bob has a contact with a Cozy URL but no email, adding this contact to
the sharing was failing. Bob's Cozy was trying to send the invitation,
but it can't create the shortcut as it requires the Xor key for this new
recipient (only available on Alice's Cozy), and it can't send an email
as the email address is not known. We have fixed this issue by making
the owner's Cozy creating the shortcut in that case.