Fix a race condition when updating io.cozy.shared
When a directory is moved from a shared folder to another shared folder,
the stack creates two jobs for updating the io.cozy.shared document
related to this directory (one for each sharing). There is a lock for
avoids the two jobs to work concurrently, but if the job for the latter
sharing was processed before the other job, the removed flag wasn't set
for the former sharing. It was due to an incorrect optimization.