Add the Servers.GetAddr method and move the tests to dynamics port attribution
When running locally it's possible to have some tests in error because
the previous test using the same port didn't have the time to close its
HTTP server, leading to a port conflict.

In order to avoid this we can use the dynamic port attribution and the
method `GetAddr` to fetch the assigned port.