Fix the maintenance list from the apps registry
When an app is present is several apps registry spaces for the same
context, it may happen that a maintenance is present in a space with a
lower priority and should not be sent to the client. The stack now
correcty handles this case by fetching all the slugs from higher
priority registries.