Fix the 'bootstrap failed' error on some CLI commands
It happens that we want to execute some cozy-stack CLI commands without
having the credentials for redis. In that case, it was failing when
initializing the logger, when trying to load the instances in debug
mode. We can ignore this error, as it not a big deal if stack doesn't
load them.