fix: Avoid to log the first redis lines on stderr
The `logger` package does need to setup `logger.debugger` in order to
know wich instances are in debug mode. One of the `debugger`
implementation (`RedisDebugger`) does need a redis client. The issue
is that once started the redis client log some lines but we don't have
setup the logger yet. This lead to have the first log lines from the
`debugger` redis client on stdout instead of syslog.

In order to fix this this commit setup the redis global logger juste
after setting up the global logger and before setting up the debugger.