Allow the Pass clients to read the instance settings
We want to implement a feature in the Pass clients (bases on bitwarden)
that requires to read the instance settings. The bitwarden protocol has
a based on OAuth2 with a fixed set of permission coded in the stack.
Changing this set of permissions would work for new clients, but for
existing clients, a migration would be required and quite complex. So,
we have prefered to modify how the permissions are checked for reading
the instance settings.