Refactor: Remove the local redis configs
The condition stopping the stack in case of conflict between the
`localOpt` variable and the `mainOpt` is setup since 2017. The
systematic setup of `mainOpt` also date of 2017.

This means that since 2017, a server with a `localOpt` can start.
We can deduce that this is feature is long gone an can be removed
safely in favor or a `Universal` client able to handle a single-node
client or a cluster client as well.