Setup the `_test` package suffix for all the tests using `testutils`
With the current way of working all the dependances
between the packages are resolved "just-in-time" via the
global variable pattern or via the `config.GetConfig().SomeStuff`
pattern. With the dependency injection all the dependances resolution
is done during the init phase and so inside the `stack.Start()` method.
This means that the stack will import all the web, the services and
the dependancies.

Before the dependency graph was mostly:
```
config <---- dependencies
       <---- stack
       <---- http
       <---- service
       <---- XXXX_test
````

This new dependency pattern lead to cycle dependency when one of the
dependency started by the stack use the package `testutils` because
this package contains an dependency to the stack.

Now it tends to:
```
stack -> http -> service -> dependencies
stack <- testutils <------- XXXX_test
(loop)
```

And this lead to a dependency cycle.

In order to solve this generic issue the `{package_name}_test` pattern
exists in order to test you package as an external package.

The proposed solution would be:
```
XXXX_test -> stack -> http -> service -> dependencies
```