Add a lock around access token endpoint
It happens that an OAuth client makes several requests in parallel for
refreshing its access token. CouchDB may create a conflict in that case.
It would be better if OAuth clients can avoid this behavior, but let's
mitigate this issue with a lock on the endpoint to avoid making the
requests to CouchDB in parallel.