Add a lock around session updates
The stack can receives several requests from a webapp in parallel. If
the session needs to be updated in CouchDB for the last_seen field, the
stack may make several requests to CouchDB in parallel, and CouchDB may
create a conflict. We can mitigate this issue by using a lock when
updating the session in CouchDB.