Add a lock around remote doctypes
When the stack receive a request for a remote doctype, it looks at the
definition of this remote doctype in CouchDB, and if not found,
downloads it from GitHub, and then puts in CouchDB. The issue is when
several calls to a remote doctypes are made in parallel. The writes in
CouchDB can result in a 202 Accepted, which then creates a conflict on
the document. We have added a lock to prevent that.