fix: Pending email with cozy-client
The Cozy client has difficulty handling field deletion via realTime. When merging, it will use the previously present field if field is not find. This allows it to update documents with requests that only select a few fields in a document. If the field is present with a value (null or empty), it will update it and the interface will react.