fix: Set correct bitwarden error (JSON) when wrong 2FA code
On Cozy Pass mobile when the login requires 2FA and when the user
enters a wrong password, then the API should return a dedicated error
message

With previous implementation, the same error would be returned on first
login attempt and on second one containing a 2FA token. This would
result on the app completing the login process but displaying an empty
vault. This behaviour is wrong but is from Bitwarden's implementation.
However with correct server implementation the correct behaviour
happens