Fix share-track jobs when out of order (bis)
It is a follow-up of #3582. When an io.cozy.file is created and modified
just after, it may happen that the share-track job for the update is
processed before the job for the creation. When this happens, the revs
tree for the io.cozy.shared wasn't correct: the revision 2 was coming
before the revision 1. The previous commit was fixing some cases, but
there is one more case that we should handle.