Improve the way we get the sharing dir
It happens that the stack needs the main directory for a sharing. For
example, when it receives a new file that is inside the shared
directory. The stack was using a referenced_by to find it, but we have
found in production that the referenced_by can have been put on the
wrong directories. We will now use the ID when we have it, and only use
the referenced_by if not.