Add a linter to disable fmt.Printf in code
Printf is often used while developping to inspect or debug code. It
happens that these statements are forgotten and commited. We can prevent
that by using a linter to disable them.

We have some cases where fmt.Printf was used to print useful stuff, like
in the CLI. They have been rewritten in fmt.Fprintf to avoid been
flagged by the linter.