Make the initial replication of a sharing faster
When a user accepts to synchronize a sharing to their Cozy, the stack
starts a replication of the io.cozy.files, and uploads the files inside
the sharing. We can make this process faster by batching calls to the
changes feed. It is particulary useful when there are lots of documents
in the io.cozy.shared database, from other sharings.