Fix a source of VFS inconsistencies
CouchDB is using ICU algorithm for views and mango requests. It can be
surprizing, and it was the source of a bug in the stack. When a
directory 'EDF' is trashed, and a directory 'EDF⚡️' was nearby, the path
of this directory can be incorrectly modified to be put in the trash.

The system test 'filepath_operations.rb' was restored from
a6a442b9075cef99667b36b7325b0a8b25eaf5b2, and updated to test this new
case.