Fix a race condition
The system test accounts_cleaning.rb has failed a couple of times on the
CI. The reason is that it tries to clean two accounts for the same
konnector at the same time, and both jobs need to recreate a work
directory and update the permission of the konnector. There was a race
condition, where it may happen that a conflict on the CouchDB document
for the permissions makes one job to fail. A retry loop will fix this
issue.