Fix email_verified_code usage
When a Cozy instance has for authentication the combo magic link + 2FA,
and its owner wants to login via the cloudery, we try to avoid a flow
with two emails (one for finding the instance domain, and the other with
the 6-digits code). To do that, we use an email_verified_code, and the
stack needs to change its behavior on the login page when this code is
present, which was not done correctly before this commit.