Fix a source of VFS inconsistency
When a new file is uploaded, and the stack tries to save the
io.cozy.files in CouchDB, it may happen that an error occurs and the
stack doesn't know if the document has been saved or not in CouchDB
(timeout, network issue, CouchDB unavailable). In that case, it is safer
to try to delete the document in CouchDB.

It is not perfect. For example, if CouchDB has accepted the document
just before becoming unavailable, the deletion will fail (CouchDB
unavailable), and it will still have an inconsistency. But, at least, it
should reduce the number of cases.