Optimize sharing for instances on the same stack
When two instances are on the same stack and there is a sharing between
for files, the stack can optimize the share-upload worker. Before this
change, the content of the file was fetched from the VFS of the source
instance, then uploaded to the destination instance, which copies it to
its VFS. Now, the stack just makes a copy from one VFS to the other.
It's a lot lighter on I/O, and should help making the share-upload
worker faster.