Fix the io.cozy.shared for some files
When some files have been in a sharing, they have an io.cozy.shared. If
they are later moved out the sharing, the io.cozy.shared remains to
synchronize this removal to the other Cozy instances. Outside of the
sharing, the files may be updated, but the revs chain of the
io.cozy.shared won't be updated. If the folder that contains those files
is moved to a sharing, the files are also added to the sharing, and
their io.cozy.shared is updated for that. What was missing was to also
update the revisions chain in that case.