Fix last-activity when OAuth clients are deleted
The last-activity is guessed from the web sessions and the OAuth
clients. But OAuth clients can be revoked/deleted, which can lose the
information. For example, if a Cozy instance has only been used with the
flagship app, and the flagship app is revoked, the last-activity would
be really wrong. We fix that by keeping a date on the instance for the
last activity for deleted OAuth clients.