Fix sharing stuck because of a missing shared ref
It happens that the job for creating an io.cozy.shared has been lost
(stack restart for example), and we need to do something to avoid having
the sharing stuck. The most efficient way to do that is to check that
the file is actually in the sharing directory, and if it is the case, to
create the missing io.cozy.shared.