Fix share-upload stuck on out-of-date io.cozy.shared
When an io.cozy.shared has not been updated with the last revision of a
file, the share-upload worker should continue its work, not getting
stuck on it. If the file was updated recently, the io.cozy.shared will
be updated, so the stack will have another try for it. If it is just
a bug that have made the io.cozy.shared out of date, blocking the worker
doesn't help. So, let's just skip the file when it happens (with a log
message to help debugging).