Fix jobs when restarting the stack
The share-upload jobs can be quite long, and they can create other jobs
(share-track). When the stack stops for a deployment, it can happen that
the share-upload tries to puhs share-track jobs after the 2 minutes
deadline, and it fails as the redis connection is closed.

The workers were modified to cancel the task context when the stack is
stopping, and the share-upload checks if the task context has been
canceled after each file upload. If the context has been canceled, it
stops there.