Fix a source of content mismatch FS errors
When the share-upload worker transfers a file from one Cozy to another,
it looks at the io.cozy.shared change feeds, and then loads the
io.cozy.files. It may happen that the two documents don't match the same
revision because of CouchDB eventual consistency, or because of a change
between the two requests. We are now aborting the share-upload in that
case to retry later.