Add a reference to sourceAccountIdentifier
When the stack executes a konnector that imports files, it checks that
a directory exists for those files. And if the directory doesn't exist,
it creates it. We have added a referencedBy on the directory for the
sourceAccountIdentifier to be able to deal with several accounts for the
same konnector, as each account should have its own directory.