Fix the myself contact at instance creation
- add the displayName
- add the index.byFamilyNameGivenNameEmailCozyUrl

In theory, there is a service from the contacts webapp that adds them,
launched via an `@vevent` trigger on the io.cozy.contacts. But, the
stack creates the myself contact before installing the contacts webapp
at instance creation. Thus, the service is not launched. The easy fix is
to just add the missing field when creating the myself contact.