Change checkbox wording at login
We now introduce the notion of data and device to emphasize the fact
that we are going to locally store data if it is checked.
Otherwise, the user won't be remembered and we will prevent to store any
personal data locally.

Note the wording itself might seem a bit verbose, but we chose to make
it explicit here. Compared to a less verbose, but also less
explicit wording, we find it legitimate here to be as explicit as
possible. We believe it is important for the user security and
empowerment, in case the login is made on an unsecured device.